move bson code to vtgate/rpcclient/gorpc
1. Move FieldType enum out of Field class.
2. Move GsonAdapters to hadoop/utils
3. Move BsonifyTest to vtgate/rpcclient/gorpc
